requirements.txt
feedparser
sumy
nltk
requests
numpy
fastapi
uvicorn
schedule
pytz
google-generativeai
yfinance
matplotlib
google-cloud-storage
beautifulsoup4
